Tested all scripts (policies + mozilla do not work)

+++ Windows/usersAndGroups.ps1
@@ -15,22 +15,44 @@ function Edit-String([string]$inputString){
     $characterArray = $inputString.ToCharArray()   
     $scrambledStringArray = $characterArray | Get-Random -Count $characterArray.Length     
     $outputString = -join $scrambledStringArray
-    $outputString = ConvertTo-SecureString $outputString
+    $outputString = ConvertTo-SecureString $outputString -AsPlainText -Force
     return $outputString
+}
+
+function Confirm-Group([string]$user, [string]$group) {
+    $inGroup = $false
+    $members = Get-LocalGroupMember -Group $group | Select-Object Name | ForEach-Object { $_.Name.split("\")[-1]  }
+    ForEach ($member in $members) {
+        if ($member -match $user) {
+            $inGroup = $true
+        }
+    }
+    return $inGroup
 }
 
 $password = Get-RandomCharacters -length 6 -characters 'abcdefghiklmnoprstuvwxyz'
 $password += Get-RandomCharacters -length 4 -characters 'ABCDEFGHKLMNOPRSTUVWXYZ'
 $password += Get-RandomCharacters -length 3 -characters '1234567890'
-$password += Get-RandomCharacters -length 3 -characters '!"§$%&/()=?}][{@#*+'
+$password += Get-RandomCharacters -length 3 -characters '!"$%&/()=?}][{@#*+'
 # 16 character password - strong as well
- 
+
+$plaintextPassword = $password
 $password = Edit-String $password # randomizes password string
 
-New-Item -Path .\OutputInput -Name "swashbucklersPassword.txt" -ItemType File -Value $password > $null # creates .txt file with password
+if (Test-Path .\OutputInput\swashbucklersPassword.txt) {
+    Remove-item .\OutputInput\swashbucklersPassword.txt
+}
+New-Item -Path .\OutputInput -ItemType File -Name "swashbucklersPassword.txt" > $null
+Add-Content -Path .\OutputInput\swashbucklersPassword.txt -Value $plaintextPassword
+
+if (Test-Path .\OutputInput\goodUsers.txt) {
+    Remove-item .\OutputInput\goodUsers.txt
+}
+New-Item -Path .\OutputInput -Name "goodUsers.txt" -ItemType File > $null # creates file for good users to be inputted
 
 Write-Host "Make sure goodUsers.txt has all of the users with all of the administrators first! (separated by lines)"
-New-Item -Path .\OutputInput -Name "goodUsers.txt" -ItemType File > $null # creates file for good users to be inputted
+Invoke-Item -Path .\OutputInput\goodUsers.txt
+
 Read-Host "Enter a character to continue.."
 
 $UserAccounts = Get-LocalUser
@@ -38,58 +60,81 @@ $UserAccounts = Get-LocalUser
 [String[]]$AdminsAllowed = (Get-Content ".\OutputInput\goodUsers.txt" | Select-Object -first $adminnumber)
 [String[]]$UsersAllowed = (Get-Content ".\OutputInput\goodUsers.txt" | Select-Object -skip $adminnumber)
 [String[]]$AllowedAccounts = $AdminsAllowed + $UsersAllowed
-$badUsers = @()
-Compare-Object -ReferenceObject $UserAccounts -DifferenceObject $AllowedAccounts -PassThru | Where-Object { $_.SideIndicator -eq '=>' } | ForEach-Object  { $badUsers += $_.InputObject }
+$badUsers = Compare-Object -ReferenceObject $UserAccounts -DifferenceObject $AllowedAccounts -PassThru
 
-
+Clear-Host
+Write-Host "`r`n`r`n----Admin Configuration----`r`n`r`n"
 ForEach ($user in $AdminsAllowed) {
     Write-Host "Adding $user [Admin account] to the Administrator group and configuring their settings."
-    Add-LocalGroupMember -Group "Administrators" -Member $user
-    Set-LocalUser -Name $user -Password $password -UserMayChangePassword $True -AccountNeverExpires $False -UserMayChangePassword $True # sets password, account can expire, user can change password
+    if ( !(Confirm-Group $user "Administrators") ) {  # If the user is not in the administrator group, add them to it
+        Add-LocalGroupMember -Group "Administrators" -Member $user
+    }
+    Set-LocalUser -Name $user -Password $password -UserMayChangePassword $true -PasswordNeverExpires $false # sets password, account can expire, user can change password
     Enable-LocalUser -Name $user
 }
 
+Write-Host "`r`n`r`n----User Configuration----`r`n`r`n"
 ForEach ($user in $UsersAllowed) {
     Write-Host "Removing $user [User account] from the Administrator group and configuring their settings."
-    Add-LocalGroupMember -Group "User" -Member $user
-    Remove-LocalGroupMember -Group "Administrators" -Member $user  # Removes user from the Administrator group
-    Set-LocalUser -Name $user -Password $password -UserMayChangePassword $True -AccountNeverExpires $False -PasswordNeverExpires $False -UserMayChangePassword $True # sets password, account can expire, user can change password
+    if ( -not (Confirm-Group $user "Users") ) {
+        Add-LocalGroupMember -Group "Users" -Member $user
+    }
+    if (Confirm-Group $user "Administrators") {
+        Remove-LocalGroupMember -Group "Administrators" -Member $user  # Removes user from the Administrator group
+    }
+    Set-LocalUser -Name $user -Password $password -UserMayChangePassword $true -PasswordNeverExpires $false  # sets password, account can expire, user can change password
     Enable-LocalUser -Name $user
 }
+
+Write-Host "`r`n`r`n----Bad Guy Configuration----`r`n`r`n"
 ForEach ($user in $badUsers) {
-    Write-Host "Disabling $user [Bad user]."
+    Write-Host "Removing and disabling $user [Bad account]."
+    if (Confirm-Group $user "Administrators") {
+        if ( !($user -match "Administrator") ) {  # Prevents the admin account from being removed
+            Remove-LocalGroupMember -Group "Administrators" -Member $user  # Removes user from the Administrator group
+        }
+    }
     Disable-LocalUser -Name $user
 }
-Write-Host "Disabling the Administrator and Guest account."
+Write-Host "`r`nDisabling the Administrator and Guest account."
 Disable-LocalUser -Name "Administrator"
 Disable-LocalUser -Name "Guest"
 # Disables admin and guest accounts
+
+
+Read-Host "Enter a character to continue.."
+Clear-Host
 
 $newuser = 'y'
 while ($newuser -eq 'y') {
     $newuser = Read-Host "Do you need to make a new user? (y/n)"
     if ($newuser -eq 'y') {
         $newusername = Read-Host "What is the user's name? (exactly)"
-        New-LocalUser -Name $newusername -Password $password -AccountNeverExpires $False -PasswordNeverExpires $False
+        New-LocalUser -Name $newusername -Password $password
+        Write-Host "User $newusername created. Welcome $newusername."
     }
 }
 
+Clear-Host
 $newgroup = 'y'
 while ($newgroup -eq 'y') {
     $newgroup = Read-Host "Do you need to make a new group? (y/n)"
-    $groupName = Read-Host "What is the group name? (exactly)"
-    New-LocalGroup -Name $groupName
-    $addUser = 'y'
-    while ($addUser -eq 'y') {
-        $addUser = Read-Host "Do you want to add a user to the group? (y/n)"
-        if ($addUser -eq 'y') {
-            $user = Read-Host "What is the name of the user you want to add? (exactly)"
-            Add-LocalGroupMember -Group $groupName -Member $user
+    if ($newgroup -eq 'y') {
+        $groupName = Read-Host "What is the group name? (exactly)"
+        New-LocalGroup -Name $groupName > $null
+        Write-Host "Group $groupName created."
+        $addUser = 'y'
+        while ($addUser -eq 'y') {
+            $addUser = Read-Host "Do you want to add a user to the group? (y/n)"
+            if ($addUser -eq 'y') {
+                $user = Read-Host "What is the name of the user you want to add? (exactly)"
+                Add-LocalGroupMember -Group $groupName -Member $user
+                Write-Host "$user added to $groupName."
+            }
         }
     }
 }
 Clear-Host
 Write-Host "Users have been audited."
-Write-Host "The password is $password. It has been written in Password.txt"
-
-
+Write-Host "The password is $plaintextPassword. It has been written in swashbucklersPassword.txt"
+Read-Host "Press any character to exit the script"
